Move subdivision filter to the geo picker component
